Member 패키지 재구성 및 Account 클래스 삭제
