--- mbting/src/main/java/kau/coop/mbting/domain/member/Gender.java
@@ -1,6 +1,0 @@
-package kau.coop.mbting.domain.member;
-
-public enum Gender {
-    MALE,
-    FEMALE,
-}
