+++ mbting/src/main/java/kau/coop/mbting/domain/member/memberinfo/Academy.java
@@ -1,4 +1,4 @@
-package kau.coop.mbting.domain.member;
+package kau.coop.mbting.domain.member.memberinfo;
 
 public enum Academy {
     한국항공대학교,
