+++ mbting/src/main/java/kau/coop/mbting/domain/member/memberinfo/Gender.java
@@ -1,0 +1,6 @@
+package kau.coop.mbting.domain.member.memberinfo;
+
+public enum Gender {
+    MALE,
+    FEMALE,
+}
